third_party/blink/web_tests/platform/generic/fast/css/content-language-mapped-to-webkit-locale-expected.txt

Test for bug 76701: map content-language to -webkit-locale.

PASS languageOfNode('x') is "\"zh-CN\""
PASS languageOfNode('y') is "\"ar\""
PASS successfullyParsed is true

TEST COMPLETE
